requirements.txt
django>=3.2.9, <3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt>=5.0.0,<5.0.1
django-cors-headers>=3.10.1,<3.10.2
Pillow>=5.3.0,<5.4.0
